// mobile/src/api/generated/models/UserReviewDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GrannFix API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserReviewDto
 */
export interface UserReviewDto {
    /**
     * 
     * @type {string}
     * @memberof UserReviewDto
     */
    offerId?: string;
    /**
     * 
     * @type {number}
     * @memberof UserReviewDto
     */
    rating?: number;
    /**
     * 
     * @type {string}
     * @memberof UserReviewDto
     */
    comment?: string;
    /**
     * 
     * @type {string}
     * @memberof UserReviewDto
     */
    reviewerId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserReviewDto
     */
    reviewerName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserReviewDto
     */
    reviewerImageUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof UserReviewDto
     */
    taskId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserReviewDto
     */
    taskTitle?: string;
    /**
     * 
     * @type {Date}
     * @memberof UserReviewDto
     */
    completedAt?: Date;
}

/**
 * Check if a given object implements the UserReviewDto interface.
 */
export function instanceOfUserReviewDto(value: object): value is UserReviewDto {
    return true;
}

export function UserReviewDtoFromJSON(json: any): UserReviewDto {
    return UserReviewDtoFromJSONTyped(json, false);
}

export function UserReviewDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserReviewDto {
    if (json == null) {
        return json;
    }
    return {
        
        'offerId': json['offerId'] == null ? undefined : json['offerId'],
        'rating': json['rating'] == null ? undefined : json['rating'],
        'comment': json['comment'] == null ? undefined : json['comment'],
        'reviewerId': json['reviewerId'] == null ? undefined : json['reviewerId'],
        'reviewerName': json['reviewerName'] == null ? undefined : json['reviewerName'],
        'reviewerImageUrl': json['reviewerImageUrl'] == null ? undefined : json['reviewerImageUrl'],
        'taskId': json['taskId'] == null ? undefined : json['taskId'],
        'taskTitle': json['taskTitle'] == null ? undefined : json['taskTitle'],
        'completedAt': json['completedAt'] == null ? undefined : (new Date(json['completedAt'])),
    };
}

export function UserReviewDtoToJSON(json: any): UserReviewDto {
    return UserReviewDtoToJSONTyped(json, false);
}

export function UserReviewDtoToJSONTyped(value?: UserReviewDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'offerId': value['offerId'],
        'rating': value['rating'],
        'comment': value['comment'],
        'reviewerId': value['reviewerId'],
        'reviewerName': value['reviewerName'],
        'reviewerImageUrl': value['reviewerImageUrl'],
        'taskId': value['taskId'],
        'taskTitle': value['taskTitle'],
        'completedAt': value['completedAt'] == null ? value['completedAt'] : value['completedAt'].toISOString(),
    };
}